pyproject.toml
[tool.poetry]
name = "crop-image"
version = "0.1.0"
description = ""
authors = ["Cologler <skyoflw@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"
typer = {extras = ["all"], version = "^0.12.5"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
